requirements.txt
asgiref==3.7.1
astroid==2.15.5
dill==0.3.6
Django==4.2.1
django-cors-headers==4.0.0
djangorestframework==3.14.0
Faker==18.9.0
isort==5.12.0
lazy-object-proxy==1.9.0
mccabe==0.7.0
platformdirs==3.5.1
pylint==2.17.4
python-dateutil==2.8.2
python-dotenv==1.0.0
pytz==2023.3
six==1.16.0
sqlparse==0.4.4
tomli==2.0.1
tomlkit==0.11.8
typing_extensions==4.6.2
validate-docbr==1.10.0
wrapt==1.15.0
